+++ src/Pusher.php
@@ -19,7 +19,7 @@ class Pusher implements LoggerAwareInterface, PusherInterface
     /**
      * @var string Version
      */
-    public static $VERSION = '7.2.2';
+    public static $VERSION = '7.2.8';
 
     /**
      * @var null|PusherCrypto
@@ -63,13 +63,6 @@ class Pusher implements LoggerAwareInterface, PusherInterface
     public function __construct(string $auth_key, string $secret, string $app_id, array $options = [], $client = null)
     {
         $this->check_compatibility();
-
-        if (!is_null($client)) {
-            $this->client = $client;
-        } else {
-            //$this->client = new \GuzzleHttp\Client();
-            $this->client = new RequestsClient();
-        }
 
         $useTLS = true;
         if (isset($options['useTLS'])) {
@@ -119,6 +112,13 @@ class Pusher implements LoggerAwareInterface, PusherInterface
                 $options['encryption_master_key_base64']
             );
             $this->crypto = new PusherCrypto($parsedKey);
+        }
+
+
+        if (!is_null($client)) {
+            $this->client = $client;
+        } else {
+            $this->client = new RequestsClient();
         }
     }
 
@@ -287,7 +287,7 @@ class Pusher implements LoggerAwareInterface, PusherInterface
         string $request_path,
         array $query_params = [],
         string $auth_version = '1.0',
-        string $auth_timestamp = null
+        ?string $auth_timestamp = null
     ): array {
         $params = [];
         $params['auth_key'] = $auth_key;
@@ -476,8 +476,8 @@ class Pusher implements LoggerAwareInterface, PusherInterface
      */
     public function sendToUser(string $user_id, string $event, $data, bool $already_encoded = false): object
     {
-      $this->validate_user_id($user_id);
-      return $this->trigger(["#server-to-user-$user_id"], $event, $data, [], $already_encoded);
+        $this->validate_user_id($user_id);
+        return $this->trigger(["#server-to-user-$user_id"], $event, $data, [], $already_encoded);
     }
 
     /**
@@ -493,8 +493,8 @@ class Pusher implements LoggerAwareInterface, PusherInterface
      */
     public function sendToUserAsync(string $user_id, string $event, $data, bool $already_encoded = false): PromiseInterface
     {
-      $this->validate_user_id($user_id);
-      return $this->triggerAsync(["#server-to-user-$user_id"], $event, $data, [], $already_encoded);
+        $this->validate_user_id($user_id);
+        return $this->triggerAsync(["#server-to-user-$user_id"], $event, $data, [], $already_encoded);
     }
 
 
@@ -725,7 +725,8 @@ class Pusher implements LoggerAwareInterface, PusherInterface
             'query' => $signature,
             'http_errors' => false,
             'headers' => $headers,
-            'base_uri' => $this->channels_url_prefix()
+            'base_uri' => $this->channels_url_prefix(),
+            'timeout' => $this->settings['timeout']
         ]);
 
         $status = $response->getStatusCode();
@@ -777,7 +778,8 @@ class Pusher implements LoggerAwareInterface, PusherInterface
                 'body' => $body,
                 'http_errors' => false,
                 'headers' => $headers,
-                'base_uri' => $this->channels_url_prefix()
+                'base_uri' => $this->channels_url_prefix(),
+                'timeout' => $this->settings['timeout']
             ]);
         } catch (ConnectException $e) {
             throw new ApiErrorException($e->getMessage());
@@ -827,7 +829,8 @@ class Pusher implements LoggerAwareInterface, PusherInterface
             'body' => $body,
             'http_errors' => false,
             'headers' => $headers,
-            'base_uri' => $this->channels_url_prefix()
+            'base_uri' => $this->channels_url_prefix(),
+            'timeout' => $this->settings['timeout'],
         ])->then(function ($response) {
             $status = $response->getStatusCode();
 
@@ -881,7 +884,7 @@ class Pusher implements LoggerAwareInterface, PusherInterface
      * @return string Json encoded authentication string.
      * @throws PusherException Throws exception if $channel is invalid or above or $socket_id is invalid
      */
-    public function authorizeChannel(string $channel, string $socket_id, string $custom_data = null): string
+    public function authorizeChannel(string $channel, string $socket_id, ?string $custom_data = null): string
     {
         $this->validate_channel($channel);
         $this->validate_socket_id($socket_id);
@@ -915,7 +918,7 @@ class Pusher implements LoggerAwareInterface, PusherInterface
         return $response;
     }
 
-     /**
+    /**
      * Convenience function for presence channel authorization.
      *
      * Equivalent to authorizeChannel($channel, $socket_id, json_encode(['user_id' => $user_id, 'user_info' => $user_info], JSON_THROW_ON_ERROR))
@@ -946,7 +949,7 @@ class Pusher implements LoggerAwareInterface, PusherInterface
     /**
      * @deprecated in favour of authorizeChannel
      */
-    public function socketAuth(string $channel, string $socket_id, string $custom_data = null): string
+    public function socketAuth(string $channel, string $socket_id, ?string $custom_data = null): string
     {
         return $this->authorizeChannel($channel, $socket_id, $custom_data);
     }
@@ -954,7 +957,7 @@ class Pusher implements LoggerAwareInterface, PusherInterface
     /**
      * @deprecated in favour of authorizeChannel
      */
-    public function socket_auth(string $channel, string $socket_id, string $custom_data = null): string
+    public function socket_auth(string $channel, string $socket_id, ?string $custom_data = null): string
     {
         return $this->authorizeChannel($channel, $socket_id, $custom_data);
     }
@@ -993,7 +996,7 @@ class Pusher implements LoggerAwareInterface, PusherInterface
         try {
             $decoded_json = json_decode($body, false, 512, JSON_THROW_ON_ERROR);
         } catch (\JsonException $e) {
-            $this->log('Unable to decrypt webhook event payload.', null, LogLevel::WARNING);
+            $this->log('Unable to decrypt webhook event payload.', [], LogLevel::WARNING);
             throw new PusherException('Data encoding error.');
         }
 
@@ -1003,12 +1006,12 @@ class Pusher implements LoggerAwareInterface, PusherInterface
                     $decryptedEvent = $this->crypto->decrypt_event($event);
 
                     if ($decryptedEvent === false) {
-                        $this->log('Unable to decrypt webhook event payload. Wrong key? Ignoring.', null, LogLevel::WARNING);
+                        $this->log('Unable to decrypt webhook event payload. Wrong key? Ignoring.', [], LogLevel::WARNING);
                         continue;
                     }
                     $decoded_events[] = $decryptedEvent;
                 } else {
-                    $this->log('Got an encrypted webhook event payload, but no master key specified. Ignoring.', null, LogLevel::WARNING);
+                    $this->log('Got an encrypted webhook event payload, but no master key specified. Ignoring.', [], LogLevel::WARNING);
                 }
             } else {
                 $decoded_events[] = $event;
@@ -1062,7 +1065,7 @@ class Pusher implements LoggerAwareInterface, PusherInterface
      */
     private function make_event(array $channels, string $event, $data, array $params = [], ?string $info = null, bool $already_encoded = false): array
     {
-      $has_encrypted_channel = false;
+        $has_encrypted_channel = false;
         foreach ($channels as $chan) {
             if (PusherCrypto::is_encrypted_channel($chan)) {
                 $has_encrypted_channel = true;
@@ -1103,12 +1106,12 @@ class Pusher implements LoggerAwareInterface, PusherInterface
         $post_params['data'] = $data_encoded;
         $channel_values = array_values($channels);
         if (count($channel_values) == 1) {
-          $post_params['channel'] = $channel_values[0];
+            $post_params['channel'] = $channel_values[0];
         } else {
-          $post_params['channels'] = $channel_values;
+            $post_params['channels'] = $channel_values;
         }
         if (!is_null($info)) {
-          $post_params['info'] = $info;
+            $post_params['info'] = $info;
         }
 
         return array_merge($post_params, $params);
@@ -1167,9 +1170,10 @@ class Pusher implements LoggerAwareInterface, PusherInterface
             $this->validate_channel($event['channel']);
             if (isset($event['socket_id'])) {
                 $this->validate_socket_id($event['socket_id']);
+                $batch[$key] = $this->make_event([$event['channel']], $event['name'], $event['data'], ['socket_id' => $event['socket_id']], $event['info'] ?? null, $already_encoded);
+            } else {
+                $batch[$key] = $this->make_event([$event['channel']], $event['name'], $event['data'], [], $event['info'] ?? null, $already_encoded);
             }
-
-            $batch[$key] = $this->make_event([$event['channel']], $event['name'], $event['data'], [], $event['info'] ?? null, $already_encoded);
         }
 
         try {